refactor: rewrite FinalPage using TypeScript
