+++ src/games/ItGame/components/Question/Question.css
@@ -43,6 +43,7 @@
   font-size: 25px;
   line-height: 30px;
   font-weight: 500;
+  text-align: center;
   background-color: var(--color-background-additional);
   color: var(--color-text-addtional);
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
