+++ src/games/ItGame/pages/FinalPage/FinalPage.tsx
@@ -3,7 +3,7 @@ import './FinalPage.css';
 import Header from '../../components/Header/Header';
 import finalEmoji from '../../images/it-game_final.svg';
 
-const FinalPage = () => {
+const FinalPage: React.FC = () => {
   return (
     <div className='page'>
       <Header />
