+++ src/games/ItGame/pages/MainPage/MainPage.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useCategories } from '../../ItGame';
-
 import Header from '../../components/Header/Header';
 import PlayingField from '../../components/PlayingField/PlayingField';
 import InfoSection from '../../components/InfoSection/InfoSection';
